+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/helpers/db/mapper/TraderNPCMapper.java
@@ -1,7 +1,7 @@
 package de.relluem94.minecraft.server.spigot.essentials.helpers.db.mapper;
 
 import de.relluem94.minecraft.server.spigot.essentials.constants.Constants;
-import de.relluem94.minecraft.server.spigot.essentials.helpers.pojo.TraderNPCEntry;
+import de.relluem94.minecraft.server.spigot.essentials.model.pojo.TraderNPCEntry;
 import de.relluem94.minecraft.server.spigot.essentials.npc.trader.TraderNPC;
 import org.bukkit.entity.Villager;
 import org.jspecify.annotations.NonNull;
@@ -10,7 +10,7 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.function.Function;
 
-import static de.relluem94.minecraft.server.spigot.essentials.constants.DatabaseMappings.*;
+import static de.relluem94.minecraft.server.spigot.essentials.constants.db.DatabaseMappings.*;
 
 public class TraderNPCMapper {
     private TraderNPCMapper() {
