+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/model/pojo/WorldEntry.java
@@ -1,4 +1,4 @@
-package de.relluem94.minecraft.server.spigot.essentials.helpers.pojo;
+package de.relluem94.minecraft.server.spigot.essentials.model.pojo;
 
 
 import lombok.Getter;
