+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/model/pojo/WorldGroupInventoryEntry.java
@@ -1,4 +1,4 @@
-package de.relluem94.minecraft.server.spigot.essentials.helpers.pojo;
+package de.relluem94.minecraft.server.spigot.essentials.model.pojo;
 
 import lombok.Getter;
 import lombok.Setter;
