+++ src/main/java/de/relluem94/minecraft/server/spigot/essentials/npc/trader/TraderNPC.java
@@ -5,8 +5,8 @@ import de.relluem94.minecraft.server.spigot.essentials.constants.Constants;
 import de.relluem94.minecraft.server.spigot.essentials.constants.ItemConstants;
 import de.relluem94.minecraft.server.spigot.essentials.helpers.ItemHelper;
 import de.relluem94.minecraft.server.spigot.essentials.helpers.ItemHelper.Rarity;
-import de.relluem94.minecraft.server.spigot.essentials.helpers.pojo.TraderNPCEntry;
-import de.relluem94.minecraft.server.spigot.essentials.npc.trader.interfaces.INPC;
+import de.relluem94.minecraft.server.spigot.essentials.interfaces.npc.INPC;
+import de.relluem94.minecraft.server.spigot.essentials.model.pojo.TraderNPCEntry;
 import org.bukkit.Material;
 import org.bukkit.entity.Villager.Profession;
 import org.bukkit.inventory.Inventory;
